Handle lines containing :line:col:
When a matched line contains a part that looks like a jump location,
like the one below, the wrong file name and location are used.
Fix this by using the same regex as grep.kak which only takes in minimal
colons.

	foobar	:2:3:	search-term

This gives up on files with colons in their filename which is probably
acceptable.